requirements.txt

basix==0.0.4
dill==0.3.5.1
GitPython==3.1.27
lightgbm==3.3.2
matplotlib==3.5.2
pandas==1.4.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
scikit-learn==1.1.1
scikit-optimize==0.9.0
sqlalchemy==1.1.2
unidecode==1.3.4
xgboost==1.6.1